Clarify language in Conflicts paragraph
I tried to make the time periods clearer.  The previous text saying "for up to two years" could have been interpreted as meaning two years is the maximum, but the conflict could expire earlier (which I don't think was the intention).  I also removed the "employer and their employees" clause, as it seemed redundant with "people in the same institution."  If that interpretation was incorrect, I can revert that change.