Add ICFP'17 to OpenTOC
As of yet, ACM has not generated an OpenTOC page for ICFP'17.  And, it
is unclear whether PACM journals will participate in OpenTOC.
However, PACMPL is Gold Open Access, so all articles will be
accessible from the ACM DL page.  And it is useful to include all
SIGPLAN events under the OpenTOC page.